multi_agents/requirements.txt
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-mcp-adapters>=0.0.1
aiohttp>=3.9.0
asyncio>=3.4.3 
python-dotenv>=1.0.1
openai>=1.0.0
langgraph>=0.0.20
requests>=2.28.0
langfuse>=2.60.4